chore(studioctl): address PR feedback
- Fold footer style tests into renderer_test.go to drop the duplicate
  in-file nolint:testpackage suppression (CodeRabbit nitpick)
- Add CHANGELOG entry for the env footer status colors
